refactoring editor component
