+++ src/components/Editor/components/Editor.js
