+++ src/components/Editor/containers/EditorContainer.js
@@ -1,52 +1,37 @@
-import Editor from "../index.js";
+import Immutable from "immutable";
+import TextEditor from "../components/Editor";
 import { connect } from "react-redux";
-import { setOutput } from "../../../actions/outputActions.js";
-import { setMostRecentProgram } from "../../../actions/userDataActions.js";
-import { setProgramDirty } from "../../../actions/programsActions.js";
-import { togglePanel } from "../../../actions/uiActions.js";
-import { CLOSED_PANEL_LEFT, OPEN_PANEL_LEFT, PANEL_SIZE } from "../../../constants";
+import { setProgramCode, setProgramDirty } from "../../../actions/programsActions.js";
 
 const mapStateToProps = state => {
-  const { mostRecentProgram } = state.userData;
+  const { uid, mostRecentProgram } = state.userData;
 
   //program data should be an object representing the most recent program
   //should have 2 keys, code (which is the code) and langauge (which is the language the code is written it)
-  const code = state.programs.getIn([mostRecentProgram, "code"], undefined);
-  const dirty = state.programs.getIn([mostRecentProgram, "dirty"], false);
-  const name = state.programs.getIn([mostRecentProgram, "name"], "untitled");
-  const language = state.programs.getIn([mostRecentProgram, "language"], "txt");
-
-  let listOfPrograms = [];
-
-  state.programs.keySeq().forEach(key => listOfPrograms.push(key));
+  // add key dirty
+  const programData = state.programs.get(mostRecentProgram, Immutable.Map()).toJS();
 
   return {
-    uid: state.userData.uid,
+    ...programData,
     mostRecentProgram,
-    code,
-    listOfPrograms,
-    screenWidth: state.ui.screenWidth,
-    screenHeight: state.ui.screenHeight,
-    dirty,
-    panelOpen: state.ui.panelOpen,
-    left: (state.ui.panelOpen ? OPEN_PANEL_LEFT : CLOSED_PANEL_LEFT) + PANEL_SIZE,
-    name,
-    language,
+    uid,
   };
 };
 
-const mapDispatchToProps = dispatch => {
+const mapDispatchToProps = (dispatch, ownProps) => {
   return {
-    setMostRecentProgram: value => dispatch(setMostRecentProgram(value)),
-    runCode: (code, language) => dispatch(setOutput(code, language)),
-    cleanCode: program => dispatch(setProgramDirty(program, false)),
-    togglePanel: () => dispatch(togglePanel()),
+    setProgramCode: (program, code) => {
+      dispatch(setProgramCode(program, code));
+    },
+    dirtyCode: program => {
+      dispatch(setProgramDirty(program, true));
+    },
   };
 };
 
-const EditorContainer = connect(
+const TextEditorContainer = connect(
   mapStateToProps,
   mapDispatchToProps,
-)(Editor);
+)(TextEditor);
 
-export default EditorContainer;
+export default TextEditorContainer;
